Update README.md
In this React Application I make Counter with Two Functionalities. which is Increment and Decrement. if we increment the value from 0 to 10 the Increment button will work properly. after getting the 10 value the increment button will be disable an if we decrement the value from 10 to 0 the decrement button will work properly. after getting the 0 value the decrement button will be disable.